Change the digest_frequency field from TINYINT to INT. JM-661
git-svn-id: http://svn.igniterealtime.org/svn/repos/wildfire/trunk@3817 b35dd754-fafc-0310-a699-88a17e54d16e

+++ src/database/wildfire_mysql.sql
@@ -267,7 +267,7 @@ CREATE TABLE pubsubSubscription (
   state               VARCHAR(15)   NOT NULL,
   deliver             TINYINT       NOT NULL,
   digest              TINYINT       NOT NULL,
-  digest_frequency    TINYINT       NOT NULL,
+  digest_frequency    INT           NOT NULL,
   expire              CHAR(15)      NULL,
   includeBody         TINYINT       NOT NULL,
   showValues          VARCHAR(30)   NULL,
